requirements.txt
absl-py
langdetect
nltk
immutabledict
openai
python-dotenv